Move player info packet handler to entity listener class
